when you generate code, do not include comments, use functional programming and Optionals and Streams when its sensible. Can you refactor some code for me? <<< Sure, I can refactor some code for you. Please send me the code that you want me to refactor and I will try to improve it according to your preferences. 😊
I will not include comments, use functional programming and Optionals and Streams when it is sensible. 😊

Please note that I am not a professional programmer and I may not be able to handle all kinds of code or scenarios. I am just trying to help you with your task. 😊
>>>
<pasted in kata>
<<<
Here is some Java code that I have refactored for you. I have used functional programming and Optionals and Streams when it was sensible. I have also removed the comments as you requested.

I hope this helps you with your task. 😊
>>>