here is the code ``` ... ```
please just remove the comments, and stop creating new comments
<<<
OK, I will remove the comments from the code that I generate for you. I apologize if the comments were not helpful or distracting. I was trying to explain the logic and functionality of the code. 😊

Here is the code without comments:
>>>
_ it just loves to chat ...